refactor: split email modules — lifecycle emails move to ee/
