+++ server/src/ee/routes/subscriptions.ts
@@ -86,14 +86,14 @@ router.post('/callback', asyncHandler(async (req, res) => {
   // Fire-and-forget subscription lifecycle emails
   const userRow = (await query('SELECT email, display_name FROM users WHERE id = $1', [userId])).rows[0];
   if (userRow?.email) {
-    const { fireSubscriptionConfirmedEmail, fireSubscriptionExpiredEmail } = await import('../../lib/emailSender.js');
+    const { fireSubscriptionConfirmedEmail, fireSubscriptionExpiredEmail } = await import('../lifecycleEmails.js');
     if (status === SubStatus.ACTIVE) {
       fireSubscriptionConfirmedEmail(userId, userRow.email, userRow.display_name, plan as import('../../lib/planGate.js').PlanTier, activeUntil);
     } else if (status === SubStatus.INACTIVE) {
       fireSubscriptionExpiredEmail(userId, userRow.email, userRow.display_name);
     }
     if (plan === Plan.PLUS && wasPro) {
-      const { fireDowngradeImpactEmail } = await import('../../lib/emailSender.js');
+      const { fireDowngradeImpactEmail } = await import('../lifecycleEmails.js');
       const [locRes, photoRes, memberRes] = await Promise.all([
         query<{ cnt: number }>('SELECT COUNT(*) as cnt FROM locations WHERE created_by = $1', [userId]),
         query<{ total: number }>('SELECT COALESCE(SUM(size), 0) as total FROM photos WHERE created_by = $1', [userId]),
